fix: Fetch schema before building and starting any dev server
Previously, tiamat graphql schema was only fetch when running dev-local, and it was fetched from the local tiamat instance. The developer had to remember to commit the schema file. Changes to the schema were not automatically propagated to deployment.

This fix ensures that schema is fetched before every build, and before starting starting any dev server. The schema is removed from source control to avoid any confusion.