Set up custom coverage tool
To instrument a function, just run
`global.markBranchHit("<funcName>", <branchIndex>)`
to mark if a branch has been hit, where <funcName> is the name of the function to be instrumented as a String, and <branchIndex> is the index of the branch in the function (starting from 0, ascending).
The branchCounts object defines the number of branches for each of the functions we instrument, so if that number is not correct, edit it.